(cisco:spark) move utility data classes out to standalone files
